LayoutTests/svg/dynamic-updates/SVGFilterPrimitiveStandardAttributes-svgdom-x-prop-expected.txt

CONSOLE MESSAGE: line 47: TypeError: 'undefined' is not an object (evaluating 'gradientElement.x.baseVal')
FAIL: Timed out waiting for notifyDone to be called
SVG 1.1 dynamic update tests

Tests dynamic updates of the 'x' property of the SVGFilterPrimitiveStandardAttributes object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


FAIL gradientElement.x.baseVal.value should be 100. Threw exception TypeError: 'undefined' is not an object (evaluating 'gradientElement.x.baseVal')
